add an optional 'noAuthCache' flag
When 'username' and 'password' flags are used without no-auth-cache, the logged in user on the machine is overridden. This is not an issue on a server, but in development environment it's very annoying. This optional flag will enable developers to access restricted repositories without changing the logged in user on their machine.